requirements.txt
bcrypt==4.3.0
blinker==1.9.0
boto3==1.39.14
botocore==1.39.14
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
marshmallow==3.20.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
s3transfer==0.13.1
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
pytesseract==0.3.10
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==20.0.0